backend/requirements/base.txt
fastapi==0.109.0
pydantic-settings==2.1.0
uvicorn==0.22.0
sqlalchemy[asyncio]==2.0.15
asyncpg==0.29.0
alembic
pyjwt
passlib[bcrypt]
websockets
